+++ ZcashLightClientKit/UIKit/Synchronizer/SDKSynchronizer.swift
@@ -529,6 +529,37 @@ public class SDKSynchronizer: Synchronizer {
         }
     }
     
+    public func rewind(_ policy: RewindPolicy) throws {
+        self.stop()
+        
+        var height: BlockHeight?
+        switch policy {
+        case .birthday:
+            let birthday = self.blockProcessor.config.walletBirthday
+            height = birthday
+            
+        case .height(let rewindHeight):
+            height = rewindHeight
+        
+        case .transaction(let tx):
+            guard let txHeight = tx.anchor else {
+                throw SynchronizerError.rewindErrorUnknownArchorHeight
+            }
+            height = txHeight
+        }
+        
+        guard let h = height else {
+            throw SynchronizerError.rewindErrorUnknownArchorHeight
+        }
+        
+        do {
+            try self.blockProcessor.rewindTo(h)
+            try self.transactionManager.handleReorg(at: h)
+        } catch {
+            throw SynchronizerError.rewindError(underlyingError: error)
+        }
+    }
+    
     // MARK: notify state
     private func notify(progress: Float, height: BlockHeight) {
         NotificationCenter.default.post(name: Notification.Name.synchronizerProgressUpdated, object: self, userInfo: [