+++ ZcashLightClientKit/zcashlc/zcashlc.h
@@ -223,6 +223,9 @@ bool zcashlc_is_valid_shielded_address(const char *address);
  */
 bool zcashlc_is_valid_transparent_address(const char *address);
 
+/**
+ * returns whether the given viewing key is valid or not
+ */
 bool zcashlc_is_valid_viewing_key(const char *key);
 
 /**